requirements.txt
Flask
numpy<2.0
tflite-runtime==2.14.0
Pillow